+++ src/main/java/org/jallen/tyrael/controller/ApplicationController.java
@@ -16,7 +16,17 @@ public class ApplicationController {
 
   @GetMapping("/applications")
   public List<Application> getAllApplications() {
-      return applicationService.getAllApplications();
+    return applicationService.getAllApplications().stream()
+      .filter(obj -> obj.getLink() != null)
+      .toList();
   }
-  
+
+  @GetMapping("/pages")
+  public List<Application> getAllPages() {
+    return applicationService.getAllApplications().stream()
+      .filter(obj -> obj.getCategory().contains("Page"))
+      .filter(obj -> obj.getLink() != null)
+      .toList();
+  }
+
 }