+++ src/main/java/org/jallen/tyrael/interfaces/Animal.java
@@ -1,0 +1,8 @@
+package org.jallen.tyrael.interfaces;
+
+public interface Animal {
+  public void eat();
+  default public void sleep() {
+    System.out.println("Animal is sleeping");
+  } 
+}