+++ src/main/java/org/jallen/tyrael/interfaces/Panda.java
@@ -1,0 +1,6 @@
+package org.jallen.tyrael.interfaces;
+
+@FunctionalInterface
+public interface Panda {
+  public String doSomething(String s); 
+}