+++ src/main/java/org/jallen/tyrael/mapper/ApplicationMapper.java
@@ -1,0 +1,5 @@
+package org.jallen.tyrael.mapper;
+
+public class ApplicationMapper {
+
+}