+++ src/main/java/org/jallen/tyrael/services/ApplicationService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.jallen.tyrael.entity.Application;
 import org.jallen.tyrael.repositories.ApplicationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,6 +15,6 @@ public class ApplicationService {
   private ApplicationRepository applicationRepository;
 
   public List<Application> getAllApplications() {
-    return applicationRepository.findAll();
+    return applicationRepository.findAll(Sort.by(Sort.Direction.ASC, "id"));
   }
 }