+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.0.1
+
+## 1.1.0 Features
+
+- `[ApplicationController]` Added web service /pages.
+- `[ApplicationMapper]` Added mapper for Application.
+- `[ReviewController]` Added controller for review purposes.
+- `[Animal]` Added animal entity for review purposes.
+- `[Dog]` Added dog entity for review purposes.
+- `[Employee]` Added dog entity for review purposes.
+
 # 1.0.0
 
 ## 1.0.0 Features
@@ -16,3 +27,4 @@ All notable changes to this project will be documented in this file.
 ## 1.0.0 Fixes
 
 - `[TyraelApplication]` Enabled cross origin for all endpoints.
+